short_url_creator/short_url_creator/requirements.txt
fastapi
"uvicorn[standard]"
sqlalchemy
"passlib[bcrypt]"
python-jose